requirements.txt
nonebot-adapter-onebot>=2.0.0
nonebot2>=2.0.0
nonebot2[httpx]
nonebot_plugin_apscheduler
pypinyin
pymongo
jieba

